# dockstore-webservice/src/main/resources/openapi.yaml
openapi: 3.0.0
info:
  description: This describes the dockstore API, a webservice that manages pairs of Docker
    images and associated metadata such as CWL documents and Dockerfiles used to
    build those images
  version: 1.7.0-alpha.2-SNAPSHOT
  title: Dockstore API
  termsOfService: TBD
  contact:
    name: Dockstore@ga4gh
    url: https://github.com/ga4gh/dockstore
    email: theglobalalliance@genomicsandhealth.org
  license:
    name: Apache License Version 2.0
    url: https://github.com/ga4gh/dockstore/blob/develop/LICENSE
...
paths:
...
  "/api/ga4gh/v1/tools/{id}":
    get:
      tags:
        - GA4GHV1
      summary: List one specific tool, acts as an anchor for self references
      description: This endpoint returns one specific tool (which has ToolVersions nested
        inside it)
      operationId: toolsIdGet
      parameters:
        - name: id
          in: path
          description: A unique identifier of the tool, scoped to this registry, for
            example `123456`
          required: true
          schema:
            type: string
      responses:
        "200":
          description: A tool.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ToolV1"
            text/plain:
              schema:
                $ref: "#/components/schemas/ToolV1"
  "/api/ga4gh/v1/tools/{id}/versions":
    get:
      tags:
        - GA4GHV1
      summary: List versions of a tool
      description: Returns all versions of the specified tool
      operationId: toolsIdVersionsGet
      parameters:
        - name: id
          in: path
          description: A unique identifier of the tool, scoped to this registry, for
            example `123456`
          required: true
          schema:
            type: string
      responses:
        "200":
          description: An array of tool versions
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ToolVersionV1"
            text/plain:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ToolVersionV1"
  "/api/ga4gh/v1/tools/{id}/versions/{version_id}":
    get:
      tags:
        - GA4GHV1
      summary: List one specific tool version, acts as an anchor for self references
      description: This endpoint returns one specific tool version
      operationId: toolsIdVersionsVersionIdGet
      parameters:
        - name: id
          in: path
          description: A unique identifier of the tool, scoped to this registry, for
            example `123456`
          required: true
          schema:
            type: string
        - name: version_id
          in: path
          description: An identifier of the tool version, scoped to this registry, for
            example `v1`
          required: true
          schema:
            type: string
      responses:
        "200":
          description: A tool version.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ToolVersionV1"
            text/plain:
              schema:
                $ref: "#/components/schemas/ToolVersionV1"
...
  /api/ga4gh/v2/extended:
    get:
      tags:
        - extendedGA4GH
      summary: List all published services, optionally filtering by subclass
      description: ""
      operationId: getPublishedServices
      parameters:
        - name: subclass
          in: query
          description: Filter by subclass
          required: false
          schema:
            type: string
      responses:
        "200":
          description: successful operation
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Service"
            text/plain:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Service"
  "/api/ga4gh/v2/extended/containers/{organization}":
    get:
      tags:
        - extendedGA4GH
      summary: List entries of an organization
      description: "This endpoint returns entries of an organization. "
      operationId: entriesOrgGet
      parameters:
        - name: organization
          in: path
          description: An organization, for example `cancercollaboratory`
          required: true
          schema:
            type: string
      responses:
        "200":
          description: An array of Tools of the input organization.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ToolV1"
            text/plain:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ToolV1"
...
  /api/ga4gh/v2/extended/tools/entry/_search:
    post:
      tags:
        - extendedGA4GH
      summary: Search the index of tools
      description: This endpoint searches the index for all published tools and workflows.
        Used by utilities that expect to talk to an elastic search endpoint
      operationId: toolsIndexSearch
      requestBody:
        content:
          application/json:
            schema:
              type: string
        description: elastic search query
        required: true
      responses:
        "200":
          description: An elastic search result.
          content:
            application/json:
              schema:
                type: string
  /api/ga4gh/v2/extended/tools/index:
    post:
      tags:
        - extendedGA4GH
      summary: Update the index of tools
      description: "This endpoint updates the index for all published tools and workflows. "
      operationId: toolsIndexGet
      responses:
        "200":
          description: An array of Tools of the input organization.
      security:
        - BEARER: []
  "/api/ga4gh/v2/extended/tools/{organization}":
    get:
      tags:
        - extendedGA4GH
      summary: List tools of an organization
      description: "This endpoint returns tools of an organization. "
      operationId: toolsOrgGet
      parameters:
        - name: organization
          in: path
          description: An organization, for example `cancercollaboratory`
          required: true
          schema:
            type: string
      responses:
        "200":
          description: An array of Tools of the input organization.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ToolV1"
            text/plain:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ToolV1"
  "/api/ga4gh/v2/extended/workflows/{organization}":
    get:
      tags:
        - extendedGA4GH
      summary: List workflows of an organization
      description: "This endpoint returns workflows of an organization. "
      operationId: workflowsOrgGet
      parameters:
        - name: organization
          in: path
          description: An organization, for example `cancercollaboratory`
          required: true
          schema:
            type: string
      responses:
        "200":
          description: An array of Tools of the input organization.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ToolV1"
            text/plain:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ToolV1"
  "/api/ga4gh/v2/extended/{id}/versions/{version_id}/{type}/tests/{relative_path}":
    post:
      tags:
        - extendedGA4GH
      summary: Annotate test JSON with information on whether it ran successfully on
        particular platforms plus metadata
      description: "Test JSON can be annotated with whether they ran correctly keyed by
        platform and associated with some metadata "
      operationId: toolsIdVersionsVersionIdTypeTestsPost
      parameters:
        - name: type
          in: path
          description: The type of the underlying descriptor. Allowable values include
            "CWL", "WDL", "NFL".
          required: true
          schema:
            type: string
        - name: id
          in: path
          description: A unique identifier of the tool, scoped to this registry, for
            example `123456`
          required: true
          schema:
            type: string
        - name: version_id
          in: path
          description: An identifier of the tool version for this particular tool registry,
            for example `v1`
          required: true
          schema:
            type: string
        - name: relative_path
          in: path
          description: A relative path to the test json as retrieved from the files
            endpoint or the tests endpoint
          required: true
          schema:
            type: string
            pattern: .+
        - name: platform
          in: query
          description: Platform to report on
          required: true
          schema:
            type: string
        - name: platform_version
          in: query
          description: Version of the platform to report on
          required: true
          schema:
            type: string
        - name: verified
          in: query
          description: Verification status, omit to delete key
          required: false
          schema:
            type: boolean
        - name: metadata
          in: query
          description: Additional information on the verification (notes, explanation)
          required: true
          schema:
            type: string
      responses:
        "200":
          description: The tool test JSON response.
          content:
            application/json:
              schema:
                type: object
                additionalProperties:
                  type: object
        "401":
          description: Credentials not provided or incorrect
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "404":
          description: The tool test cannot be found to annotate.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
      security:
        - BEARER: []
  "/api/ga4gh/v2/extended/{serviceId}":
    get:
      tags:
        - extendedGA4GH
      summary: Fetches the specified service version
      description: ""
      operationId: getService
      parameters:
        - name: serviceId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        "200":
          description: successful operation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Service"
            text/plain:
              schema:
                $ref: "#/components/schemas/Service"
  "/api/ga4gh/v2/extended/{serviceId}/versions":
    get:
      tags:
        - extendedGA4GH
      summary: Gets all versions of the specified version
      description: ""
      operationId: getServiceVersions
      parameters:
        - name: serviceId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        "200":
          description: successful operation
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ServiceVersion"
            text/plain:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ServiceVersion"
  "/api/ga4gh/v2/extended/{serviceId}/versions/{versionId}":
    get:
      tags:
        - extendedGA4GH
      summary: Gets the specified service version
      description: ""
      operationId: getServiceVersion
      parameters:
        - name: serviceId
          in: path
          required: true
          schema:
            type: integer
            format: int64
        - name: versionId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        "200":
          description: successful operation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ServiceVersion"
            text/plain:
              schema:
                $ref: "#/components/schemas/ServiceVersion"
...
  "/api/ga4gh/v2/tools/{id}":
    get:
      tags:
        - GA4GH
      summary: List one specific tool, acts as an anchor for self references
      description: This endpoint returns one specific tool (which has ToolVersions nested
        inside it)
      operationId: toolsIdGet
      parameters:
        - name: id
          in: path
          description: A unique identifier of the tool, scoped to this registry, for
            example `123456`
          required: true
          schema:
            type: string
      responses:
        "200":
          description: A tool.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Tool"
            text/plain:
              schema:
                $ref: "#/components/schemas/Tool"
        "404":
          description: The tool can not be found.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
            text/plain:
              schema:
                $ref: "#/components/schemas/Error"
  "/api/ga4gh/v2/tools/{id}/versions":
    get:
      tags:
        - GA4GH
      summary: List versions of a tool
      description: Returns all versions of the specified tool
      operationId: toolsIdVersionsGet
      parameters:
        - name: id
          in: path
          description: A unique identifier of the tool, scoped to this registry, for
            example `123456`
          required: true
          schema:
            type: string
      responses:
        "200":
          description: An array of tool versions
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ToolVersion"
            text/plain:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ToolVersion"
  "/api/ga4gh/v2/tools/{id}/versions/{version_id}":
    get:
      tags:
        - GA4GH
      summary: List one specific tool version, acts as an anchor for self references
      description: This endpoint returns one specific tool version
      operationId: toolsIdVersionsVersionIdGet
      parameters:
        - name: id
          in: path
          description: A unique identifier of the tool, scoped to this registry, for
            example `123456`
          required: true
          schema:
            type: string
        - name: version_id
          in: path
          description: An identifier of the tool version, scoped to this registry, for
            example `v1`
          required: true
          schema:
            type: string
      responses:
        "200":
          description: A tool version.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ToolVersion"
            text/plain:
              schema:
                $ref: "#/components/schemas/ToolVersion"
        "404":
          description: The tool can not be found.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
            text/plain:
              schema:
                $ref: "#/components/schemas/Error"
...
  "/containers/path/tool/{repository}":
    get:
      tags:
        - containers
      summary: Get a tool by the specific tool path
      description: Requires full path (including tool name if applicable).
      operationId: getContainerByToolPath
      parameters:
        - name: repository
          in: path
          description: repository path
          required: true
          schema:
            type: string
        - name: include
          in: query
          description: "Comma-delimited list of fields to include: validations"
          required: false
          schema:
            type: string
      responses:
        "200":
          description: successful operation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DockstoreTool"
      security:
        - BEARER: []
...
  "/containers/{containerId}":
    get:
      tags:
        - containers
      summary: Retrieve a tool.
      description: This is one of the few endpoints that returns the user object with
        populated properties (minus the userProfiles property)
      operationId: getContainer
      parameters:
        - name: containerId
          in: path
          description: Tool ID
          required: true
          schema:
            type: integer
            format: int64
        - name: include
          in: query
          description: "Comma-delimited list of fields to include: validations"
          required: false
          schema:
            type: string
      responses:
        "200":
          description: successful operation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DockstoreTool"
      security:
        - BEARER: []
    put:
      tags:
        - containers
      summary: Update the tool with the given tool.
      description: Updates default descriptor paths, default Docker paths, default test
        parameter paths, git url, and default version. Also updates tool
        maintainer email, and private access for manual tools.
      operationId: updateContainer
      parameters:
        - name: containerId
          in: path
          description: Tool to modify.
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        $ref: "#/components/requestBodies/DockstoreTool"
      responses:
        "200":
          description: successful operation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DockstoreTool"
      security:
        - BEARER: []
    delete:
      tags:
        - containers
      summary: Delete a tool.
      description: ""
      operationId: deleteContainer
      parameters:
        - name: containerId
          in: path
          description: Tool id to delete
          required: true
          schema:
            type: integer
            format: int64
      responses:
        "400":
          description: "Invalid "
      security:
        - BEARER: []
...
  /organizations:
    get:
      tags:
        - organizations
      summary: List all available organizations.
      description: NO Authentication
      operationId: getApprovedOrganizations
      responses:
        "200":
          description: successful operation
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Organization"
    post:
      tags:
        - organizations
      summary: Create an organization.
      description: Organization requires approval by an admin before being made public.
      operationId: createOrganization
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Organization"
        description: Organization to register.
        required: true
      responses:
        "200":
          description: successful operation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Organization"
      security:
        - BEARER: []
...
  "/organizations/{organizationId}":
    get:
      tags:
        - organizations
      summary: Retrieves an organization by ID.
      description: Does not require authentication for approved organizations,
        authentication can be provided for unapproved organizations
      operationId: getOrganizationById
      parameters:
        - name: organizationId
          in: path
          description: Organization ID.
          required: true
          schema:
            type: integer
            format: int64
      responses:
        "200":
          description: successful operation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Organization"
      security:
        - BEARER: []
    put:
      tags:
        - organizations
      summary: Update an organization.
      description: Currently only name, display name, description, topic, email, link,
        avatarUrl, and location can be updated.
      operationId: updateOrganization
      parameters:
        - name: organizationId
          in: path
          description: Organization ID.
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Organization"
        description: Organization to update with.
        required: true
      responses:
        "200":
          description: successful operation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Organization"
      security:
        - BEARER: []
...
  /services:
    get:
      tags:
        - services
      summary: List all published services, optionally filtering by subclass
      description: ""
      operationId: getPublishedServices
      parameters:
        - name: subclass
          in: query
          description: Filter by subclass
          required: false
          schema:
            type: string
      responses:
        "200":
          description: successful operation
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Service"
  "/services/{serviceId}":
    get:
      tags:
        - services
      summary: Fetches the specified service version
      description: ""
      operationId: getService
      parameters:
        - name: serviceId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        "200":
          description: successful operation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Service"
  "/services/{serviceId}/versions":
    get:
      tags:
        - services
      summary: Gets all versions of the specified version
      description: ""
      operationId: getServiceVersions
      parameters:
        - name: serviceId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        "200":
          description: successful operation
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ServiceVersion"
...
servers:
  - url: https://dockstore.org/api
components:
...
  securitySchemes:
    BEARER:
      type: apiKey
      name: Authorization
      in: header
  schemas:
...
    Error:
      type: object
      required:
        - code
      properties:
        code:
          type: integer
          format: int32
        message:
          type: string
...
    ToolV1:
      type: object
      properties:
        url:
          type: string
        id:
          type: string
        organization:
          type: string
        toolname:
          type: string
        toolclass:
          $ref: "#/components/schemas/ToolClass"
        description:
          type: string
        author:
          type: string
        meta-version:
          type: string
        contains:
          type: array
          items:
            type: string
        verified:
          type: boolean
        verified-source:
          type: string
        signed:
          type: boolean
        versions:
          type: array
          items:
            $ref: "#/components/schemas/ToolVersionV1"
      description: A tool (or described tool) describes one pairing of a tool as described
        in a descriptor file (which potentially describes multiple tools) and a
        Docker image.
...